+++ src/partest/scala/tools/partest/nest/FileManager.scala
@@ -12,17 +12,9 @@ import java.io.{File, FilenameFilter, IOException, StringWriter,
                 FileInputStream, FileOutputStream, BufferedReader,
                 FileReader, PrintWriter, FileWriter}
 import java.net.URI
-import scala.tools.nsc.io.Directory
+import scala.tools.nsc.io.{ Path, Directory }
 
-trait FileManager {
-
-  def basename(name: String): String = {
-    val inx = name.lastIndexOf(".")
-    if (inx < 0) name else name.substring(0, inx)
-  }
-
-  def deleteRecursive(dir: File) { Directory(dir).deleteRecursively() }
-  
+trait FileManager {  
   /**
    * Compares two files using a Java implementation of the GNU diff
    * available at http://www.bmsi.com/java/#diff.
@@ -32,19 +24,12 @@ trait FileManager {
    * @return the text difference between the compared files
    */
   def compareFiles(f1: File, f2: File): String = {
-    var res = ""
-    try {
-      val diffWriter = new StringWriter
-      val args = Array(f1.getCanonicalPath(), f2.getCanonicalPath())
-      DiffPrint.doDiff(args, diffWriter)
-      res = diffWriter.toString
-      if (res startsWith "No")
-        res = ""
-    } catch {
-      case e: IOException =>
-        e.printStackTrace()
-    }
-    res
+    val diffWriter = new StringWriter
+    val args = Array(f1.getCanonicalPath(), f2.getCanonicalPath())
+    
+    DiffPrint.doDiff(args, diffWriter)
+    val res = diffWriter.toString
+    if (res startsWith "No") "" else res
   }
 
   var JAVACMD: String
@@ -60,10 +45,9 @@ trait FileManager {
   var showLog = false
   var failed = false
 
-  var SCALAC_OPTS = System.getProperty("scalatest.scalac_opts", "-deprecation")
-  var JAVA_OPTS   = System.getProperty("scalatest.java_opts", "")
-
-  var timeout = "1200000"
+  var SCALAC_OPTS = PartestDefaults.scalacOpts
+  var JAVA_OPTS   = PartestDefaults.javaOpts
+  var timeout     = PartestDefaults.timeout
 
   def getLogFile(dir: File, fileBase: String, kind: String): LogFile =
     new LogFile(dir, fileBase + "-" + kind + ".log")
@@ -74,47 +58,34 @@ trait FileManager {
     getLogFile(dir, fileBase, kind)
   }
 
-  def logFileExists(file: File, kind: String): Boolean = {
-    val logFile = getLogFile(file, kind)
-    logFile.exists && logFile.canRead
-  }
+  def logFileExists(file: File, kind: String) =
+    getLogFile(file, kind).canRead
   
-  def overwriteFileWith(dest: File, file: File): Boolean =
-    if (dest.exists && dest.isFile)
-        copyFile(file, dest)
-    else
-      false
+  def overwriteFileWith(dest: File, file: File) =
+    dest.isFile && copyFile(file, dest)
   
   def copyFile(from: File, to: File): Boolean =
     try {
-      val fromReader = new BufferedReader(new FileReader(from))
-      val toWriter = new PrintWriter(new FileWriter(to))
-      
-      new StreamAppender(fromReader, toWriter).run()
-      
-      fromReader.close()
-      toWriter.close()
+      val appender = StreamAppender(from, to)
+      appender.run()
+      appender.closeAll()
       true
-    } catch {
-      case _:IOException => false
+    }
+    catch {
+      case _: IOException => false
     }
 
   def mapFile(file: File, suffix: String, dir: File, replace: String => String) {
     val tmpFile = File.createTempFile("tmp", suffix, dir) // prefix required by API
-    val fileReader = new BufferedReader(new FileReader(file))
-    val tmpFilePrinter = new PrintWriter(new FileWriter(tmpFile))
-    val appender = new StreamAppender(fileReader, tmpFilePrinter)
     
+    val appender = StreamAppender(file, tmpFile)    
     appender.runAndMap(replace)
-    
-    fileReader.close()
-    tmpFilePrinter.close()
+    appender.closeAll()
 
-    val tmpFileReader = new BufferedReader(new FileReader(tmpFile))
-    val filePrinter= new PrintWriter(new FileWriter(file), true)
-    (new StreamAppender(tmpFileReader, filePrinter)).run
-    tmpFileReader.close()
-    filePrinter.close()
+    val appender2 = StreamAppender(tmpFile, file)
+    appender2.run()
+    appender2.closeAll()
+    
     tmpFile.delete()
   }
 }
